+++ HISTORY.rst
@@ -2,6 +2,11 @@
 History
 =======
 
+Unreleased
+----------
+
+* Added principal logarithm `MultiVector.log()` for simple rotors with scalar and bivector parts.
+
 0.1.0 (2023-08-12)
 ------------------